Expose GpuMetric create APIs so that they can be used in non-GpuExec
