+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/BasicColumnarWriteStatsTracker.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2023, NVIDIA CORPORATION.
+ * Copyright (c) 2019-2025, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,6 +21,7 @@ import java.nio.charset.StandardCharsets
 
 import scala.collection.mutable
 
+import com.nvidia.spark.rapids.GpuMetric
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.{FileSystem, Path}
 
@@ -28,7 +29,7 @@ import org.apache.spark.{SparkContext, TaskContext}
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.execution.SQLExecution
 import org.apache.spark.sql.execution.datasources.WriteTaskStats
-import org.apache.spark.sql.execution.metric.{SQLMetric, SQLMetrics}
+import org.apache.spark.sql.execution.metric.SQLMetrics
 import org.apache.spark.sql.rapids.BasicColumnarWriteJobStatsTracker._
 import org.apache.spark.sql.vectorized.ColumnarBatch
 import org.apache.spark.util.SerializableConfiguration
@@ -52,7 +53,7 @@ case class BasicColumnarWriteTaskStats(
  */
 class BasicColumnarWriteTaskStatsTracker(
     hadoopConf: Configuration,
-    taskCommitTimeMetric: Option[SQLMetric])
+    taskCommitTimeMetric: Option[GpuMetric])
     extends ColumnarWriteTaskStatsTracker with Logging {
   private[this] var numPartitions: Int = 0
   private[this] var numFiles: Int = 0
@@ -184,13 +185,13 @@ class BasicColumnarWriteTaskStatsTracker(
  */
 class BasicColumnarWriteJobStatsTracker(
     serializableHadoopConf: SerializableConfiguration,
-    @transient val driverSideMetrics: Map[String, SQLMetric],
-    taskCommitTimeMetric: SQLMetric)
+    @transient val driverSideMetrics: Map[String, GpuMetric],
+    taskCommitTimeMetric: GpuMetric)
   extends ColumnarWriteJobStatsTracker {
 
   def this(
       serializableHadoopConf: SerializableConfiguration,
-      metrics: Map[String, SQLMetric]) = {
+      metrics: Map[String, GpuMetric]) = {
     this(serializableHadoopConf, metrics - TASK_COMMIT_TIME, metrics(TASK_COMMIT_TIME))
   }
 
@@ -221,7 +222,8 @@ class BasicColumnarWriteJobStatsTracker(
     driverSideMetrics(BasicColumnarWriteJobStatsTracker.NUM_PARTS_KEY).add(numPartitions)
 
     val executionId = sparkContext.getLocalProperty(SQLExecution.EXECUTION_ID_KEY)
-    SQLMetrics.postDriverMetricUpdates(sparkContext, executionId, driverSideMetrics.values.toList)
+    SQLMetrics.postDriverMetricUpdates(sparkContext, executionId,
+      GpuMetric.unwrap(driverSideMetrics).values.toList)
   }
 }
 
@@ -235,15 +237,20 @@ object BasicColumnarWriteJobStatsTracker {
   /** XAttr key of the data length header added in HADOOP-17414. */
   val FILE_LENGTH_XATTR = "header.x-hadoop-s3a-magic-data-length"
 
-  def metrics: Map[String, SQLMetric] = {
-    val sparkContext = SparkContext.getActive.get
+  def metrics: Map[String, GpuMetric] = {
     Map(
-      NUM_FILES_KEY -> SQLMetrics.createMetric(sparkContext, "number of written files"),
-      NUM_OUTPUT_BYTES_KEY -> SQLMetrics.createSizeMetric(sparkContext, "written output"),
-      NUM_OUTPUT_ROWS_KEY -> SQLMetrics.createMetric(sparkContext, "number of output rows"),
-      NUM_PARTS_KEY -> SQLMetrics.createMetric(sparkContext, "number of dynamic part"),
-      TASK_COMMIT_TIME -> SQLMetrics.createTimingMetric(sparkContext, "task commit time"),
-      JOB_COMMIT_TIME -> SQLMetrics.createTimingMetric(sparkContext, "job commit time")
+      NUM_FILES_KEY -> GpuMetric.create(GpuMetric.ESSENTIAL_LEVEL,
+        "number of written files"),
+      NUM_OUTPUT_BYTES_KEY -> GpuMetric.createSizeMetric(GpuMetric.ESSENTIAL_LEVEL,
+        "written output"),
+      NUM_OUTPUT_ROWS_KEY -> GpuMetric.create(GpuMetric.ESSENTIAL_LEVEL,
+        "number of output rows"),
+      NUM_PARTS_KEY -> GpuMetric.create(GpuMetric.ESSENTIAL_LEVEL,
+        "number of dynamic part"),
+      TASK_COMMIT_TIME -> GpuMetric.createTimingMetric(GpuMetric.ESSENTIAL_LEVEL,
+        "task commit time"),
+      JOB_COMMIT_TIME -> GpuMetric.createTimingMetric(GpuMetric.ESSENTIAL_LEVEL,
+        "job commit time")
     )
   }
 }
