+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/GpuWriteStatsTracker.scala
@@ -16,14 +16,9 @@
 
 package org.apache.spark.sql.rapids
 
-import java.util.concurrent.TimeUnit
-
-import com.nvidia.spark.rapids.GpuDataWritingCommand
-import com.nvidia.spark.rapids.GpuMetric
+import com.nvidia.spark.rapids.{GpuDataWritingCommand, GpuMetric}
 import org.apache.hadoop.conf.Configuration
 
-import org.apache.spark.SparkContext
-import org.apache.spark.sql.execution.metric.{SQLMetric, SQLMetrics}
 import org.apache.spark.sql.rapids.BasicColumnarWriteJobStatsTracker.TASK_COMMIT_TIME
 import org.apache.spark.util.SerializableConfiguration
 
@@ -33,7 +28,7 @@ import org.apache.spark.util.SerializableConfiguration
  */
 class GpuWriteTaskStatsTracker(
     hadoopConf: Configuration,
-    taskMetrics: Map[String, SQLMetric])
+    taskMetrics: Map[String, GpuMetric])
     extends BasicColumnarWriteTaskStatsTracker(hadoopConf, taskMetrics.get(TASK_COMMIT_TIME)) {
   def addGpuTime(nanos: Long): Unit = {
     taskMetrics(GpuWriteJobStatsTracker.GPU_TIME_KEY) += nanos
@@ -43,8 +38,16 @@ class GpuWriteTaskStatsTracker(
     taskMetrics(GpuWriteJobStatsTracker.WRITE_TIME_KEY) += nanos
   }
 
-  def setAsyncWriteThrottleTimes(avgNs: Double, minNs: Long, maxNs: Long): Unit = {
-    taskMetrics(GpuWriteJobStatsTracker.ASYNC_WRITE_AVG_THROTTLE_TIME_KEY).set(avgNs.toLong)
+  def setAsyncWriteThrottleTimes(numTasks: Int, accumulatedThrottleTimeNs: Long, minNs: Long,
+      maxNs: Long): Unit = {
+    val avg = if (numTasks > 0) {
+      accumulatedThrottleTimeNs.toDouble / numTasks
+    } else {
+      0
+    }
+    taskMetrics(GpuWriteJobStatsTracker.ASYNC_WRITE_TOTAL_THROTTLE_TIME_KEY).set(
+      accumulatedThrottleTimeNs)
+    taskMetrics(GpuWriteJobStatsTracker.ASYNC_WRITE_AVG_THROTTLE_TIME_KEY).set(avg.toLong)
     taskMetrics(GpuWriteJobStatsTracker.ASYNC_WRITE_MIN_THROTTLE_TIME_KEY).set(minNs)
     taskMetrics(GpuWriteJobStatsTracker.ASYNC_WRITE_MAX_THROTTLE_TIME_KEY).set(maxNs)
   }
@@ -58,8 +61,8 @@ class GpuWriteTaskStatsTracker(
  */
 class GpuWriteJobStatsTracker(
     serializableHadoopConf: SerializableConfiguration,
-    @transient driverSideMetrics: Map[String, SQLMetric],
-    taskMetrics: Map[String, SQLMetric])
+    @transient driverSideMetrics: Map[String, GpuMetric],
+    taskMetrics: Map[String, GpuMetric])
     extends BasicColumnarWriteJobStatsTracker(serializableHadoopConf, driverSideMetrics) {
   override def newTaskInstance(): ColumnarWriteTaskStatsTracker = {
     new GpuWriteTaskStatsTracker(serializableHadoopConf.value, taskMetrics)
@@ -69,25 +72,27 @@ class GpuWriteJobStatsTracker(
 object GpuWriteJobStatsTracker {
   val GPU_TIME_KEY = "gpuTime"
   val WRITE_TIME_KEY = "writeTime"
+  val ASYNC_WRITE_TOTAL_THROTTLE_TIME_KEY = "asyncWriteTotalThrottleTime"
   val ASYNC_WRITE_AVG_THROTTLE_TIME_KEY = "asyncWriteAvgThrottleTime"
   val ASYNC_WRITE_MIN_THROTTLE_TIME_KEY = "asyncWriteMinThrottleTime"
   val ASYNC_WRITE_MAX_THROTTLE_TIME_KEY = "asyncWriteMaxThrottleTime"
 
-  // TODO: make them GpuMetrics with metricsLevel
-  def basicMetrics: Map[String, SQLMetric] = BasicColumnarWriteJobStatsTracker.metrics
+  def basicMetrics: Map[String, GpuMetric] = BasicColumnarWriteJobStatsTracker.metrics
 
-  def taskMetrics: Map[String, SQLMetric] = {
-    val sparkContext = SparkContext.getActive.get
+  def taskMetrics: Map[String, GpuMetric] = {
     Map(
-      GPU_TIME_KEY -> SQLMetrics.createNanoTimingMetric(sparkContext, "GPU time"),
-      WRITE_TIME_KEY -> SQLMetrics.createNanoTimingMetric(sparkContext, "write time"),
+      GPU_TIME_KEY -> GpuMetric.createNanoTimingMetric(GpuMetric.ESSENTIAL_LEVEL, "GPU time"),
+      WRITE_TIME_KEY -> GpuMetric.createNanoTimingMetric(GpuMetric.ESSENTIAL_LEVEL,
+        "write time"),
       TASK_COMMIT_TIME -> basicMetrics(TASK_COMMIT_TIME),
-      ASYNC_WRITE_AVG_THROTTLE_TIME_KEY -> SQLMetrics.createNanoTimingMetric(sparkContext,
-        "async write avg throttle time"),
-      ASYNC_WRITE_MIN_THROTTLE_TIME_KEY -> SQLMetrics.createNanoTimingMetric(sparkContext,
-        "async write min throttle time"),
-      ASYNC_WRITE_MAX_THROTTLE_TIME_KEY -> SQLMetrics.createNanoTimingMetric(sparkContext,
-        "async write max throttle time")
+      ASYNC_WRITE_TOTAL_THROTTLE_TIME_KEY -> GpuMetric.createNanoTimingMetric(
+        GpuMetric.DEBUG_LEVEL, "total throttle time"),
+      ASYNC_WRITE_AVG_THROTTLE_TIME_KEY -> GpuMetric.createNanoTimingMetric(
+        GpuMetric.DEBUG_LEVEL, "avg throttle time per async write"),
+      ASYNC_WRITE_MIN_THROTTLE_TIME_KEY -> GpuMetric.createNanoTimingMetric(
+        GpuMetric.DEBUG_LEVEL, "min throttle time per async write"),
+      ASYNC_WRITE_MAX_THROTTLE_TIME_KEY -> GpuMetric.createNanoTimingMetric(
+        GpuMetric.DEBUG_LEVEL, "max throttle time per async write")
     )
   }
 
@@ -96,7 +101,7 @@ object GpuWriteJobStatsTracker {
     new GpuWriteJobStatsTracker(serializableHadoopConf, command.basicMetrics, command.taskMetrics)
 
   def apply(serializableHadoopConf: SerializableConfiguration,
-      basicMetrics: Map[String, SQLMetric],
-      taskMetrics: Map[String, SQLMetric]): GpuWriteJobStatsTracker = 
+      basicMetrics: Map[String, GpuMetric],
+      taskMetrics: Map[String, GpuMetric]): GpuWriteJobStatsTracker =
     new GpuWriteJobStatsTracker(serializableHadoopConf, basicMetrics, taskMetrics)
 }
